Feat: add etch-a-sketch layout
the layout is aiming to imitate the looks of a
vintage "magic Etch A Sketch Screen"
the top section with the game title is a single div, spanning
the full width of the game container.
the left border, center and right border share a row container.
each of the three parts will be in column layouts.
the center div includes a screen div and the bottom border div.